Satisfy clippy

+++ rclrs/src/context.rs
@@ -217,7 +217,7 @@ impl InitOptions {
             // other than None. When the user asks for None, that is equivalent
             // to the default value in rcl_init_options.
             if let Some(domain_id) = self.domain_id {
-                rcl_init_options_set_domain_id(&mut rcl_init_options, domain_id as usize);
+                rcl_init_options_set_domain_id(&mut rcl_init_options, domain_id);
             }
             Ok(rcl_init_options)
         }
